# public/swagger/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": " stores API",
        "version": "0.1"
    },
    "servers": [
        {
            "url": "http://localhost:8080",
            "description": "Serveur local"
        },
        {
            "url": "http://localhost:8080",
            "description": "Serveur local"
        }
    ],
    "paths": {
...
        "/user/create": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Créer un utilisateur",
                "description": "Permet de créer un utilisateur.",
                "operationId": "4f6ec1930302582f147f55a61239de55",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "description": "Email de l'utilisateur",
                                        "type": "string",
                                        "format": "email"
                                    },
                                    "password": {
                                        "description": "Mot de passe de l'utilisateur",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Utilisateur créé avec succès"
                    },
                    "400": {
                        "description": "Données invalides"
                    }
                }
            }
        },
        "/user/login": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "permet de se connecter",
                "description": "Se connecter et récupérer un token.",
                "operationId": "bb52867672b75505465de49d5ca37c6e",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Utilisateur connecté avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Champs manquants ou mauvais identifiants"
                    },
                    "500": {
                        "description": "Erreur lors de la connexion de l'utilisateur"
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "BearerToken": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}